Extract wrk driver to separate project
