+++ pom.xml
@@ -30,12 +30,6 @@
             <version>4.12</version>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>com.xebialabs.restito</groupId>
-            <artifactId>restito</artifactId>
-            <version>0.9.3</version>
-            <scope>test</scope>
-        </dependency>
     </dependencies>
 
     <build>
