--- src/main/java/com/github/loadtest4j/loadtest4j/drivers/wrk/Command.java
@@ -1,23 +1,0 @@
-package com.github.loadtest4j.loadtest4j.drivers.wrk;
-
-import java.util.Collections;
-import java.util.List;
-
-class Command {
-
-    private final List<String> arguments;
-    private final String launchPath;
-
-    Command(List<String> arguments, String launchPath) {
-        this.arguments = arguments;
-        this.launchPath = launchPath;
-    }
-
-    protected List<String> getArguments() {
-        return Collections.unmodifiableList(arguments);
-    }
-
-    protected String getLaunchPath() {
-        return launchPath;
-    }
-}
