--- src/test/java/com/github/loadtest4j/loadtest4j/DriverFactoryTest.java
@@ -1,20 +1,0 @@
-package com.github.loadtest4j.loadtest4j;
-
-import org.junit.Test;
-
-public abstract class DriverFactoryTest {
-    protected abstract DriverFactory sut();
-
-    @Test
-    public abstract void testGetMandatoryProperties();
-
-    @Test(expected = UnsupportedOperationException.class)
-    public void testGetMandatoryPropertiesIsImmutable() {
-        final DriverFactory sut = sut();
-
-        sut.getMandatoryProperties().add("foobarbaz123");
-    }
-
-    @Test
-    public abstract void testCreate();
-}
